upload: Move methods into zerver.lib.upload from .base.
